+++ mui/src/components/ColumnChart/types.ts
@@ -1,0 +1,5 @@
+export interface ColumnData {
+  value: number,
+  label?: string,
+  legend: string
+}