+++ mui/src/pages/Dashboard/Dashboard.tsx
@@ -1,20 +1,17 @@
 import StatusIndicator from "@/features/StatusIndicator";
-import ColumnChart from "@/features/ColumnChart";
 import styles from './Dashboard.module.css';
-
-const INITIAL_DATA = [
-  { column_name: "TX", value: 100 },
-  { column_name: "RX", value: 150 }
-];
+import PacketsColumnChart from "@/features/PacketsColumnChart";
+import Header from "@/components/Header/Header";
 
 function Dashboard() {
   return (
-    <div className={styles.page}>
-      <StatusIndicator />
-      <ColumnChart
-        data={INITIAL_DATA}
-      />
-    </div>
+    <>
+      <Header />
+      <div className={styles.page}>
+        <StatusIndicator />
+        <PacketsColumnChart/>
+      </div>
+    </>
   );
 }